Refactor benchmark configuration handling
